pyproject.toml
[project]
name = "liveethicsdata"
version = "1.2.3"
description = "A package for company ethics data aggregation."
authors = [
    { name = "Kai Sereni", email = "kaiserenior@gmail.com" }
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "beautifulsoup4==4.13.3",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "exceptiongroup==1.2.2",
    "gcloud==0.18.3",
    "google-auth==2.38.0",
    "google-genai==1.3.0",
    "googleapis-common-protos==1.68.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "oauth2client==4.1.3",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pydantic==2.11.0a2",
    "pydantic_core==2.29.0",
    "pyparsing==3.2.1",
    "python-dotenv",
    "requests==2.32.3",
    "rsa==4.9",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "typing_extensions==4.12.2",
    "urllib3==2.3.0",
    "websockets==14.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["liveethicsdata"]